refactor(#273): split ChatPage.tsx into sub-components and hooks
Extract the 1515-line ChatPage.tsx into focused modules:
- chatPageTypes.ts: shared types, constants, and utility functions
- useChatPageState.ts: all state management, hooks composition, and effects
- chatPageHandlers.ts: all handler/callback logic (commands, workflows, plan, rewind, etc.)
- CostDialog.tsx: session cost dialog sub-component
- LinearIssueBar.tsx: linear issue context bar sub-component
- CommandTipBanner.tsx: dismissable command tip banner sub-component

ChatPage.tsx is now a 291-line orchestrator that composes state, handlers,
and renders sub-components.
